requirements.txt
torch
torchvision
numpy
tqdm
Pillow
einops
pyyaml
matplotlib
scipy
